% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbankeR.R
\docType{package}
\name{openbankeR}
\alias{openbankeR}
\title{openbankeR: a package for querying the UK Open Banking (Open Data) API}
\description{
The openbankeR package provides two main features:
\itemize{
  \item Create an R client for the API
  \item Extract raw data from the API using convenient queries
}
}
\section{Available functions}{

The available functions are:
\itemize{
  \item \code{\link{CreateOpenBankingClient}}: Creates a client for querying the API
}
}

